Unify `performRead*` variants with one taking `struct msghdr`
Summary:
This is a plumbing refactor that is intended to be a no-op in terms of functionality (except for the `TAsyncSocketIntercepted` and `AsyncKTLSSocket` enhancements noted below), and a no-op in terms of performance.

When trying to make `ReadAncillaryDataCallback` safer in D43588338, I ran across the issue of _how_ to plumb the ancillary data (`msg.msg_control*`) from `AsyncSocket::performReadInternal` out to `processNormalRead`.

I initially tried adding the data to `ReadStatus` but simpkins pointed out that the RVO/NVRO setup became fragile, or outright broken by the resulting more complex return paths in `performReadInternal`. He then observed that passing `struct ::msghdr&` from the caller could overall simplify the code.

After coding up the PoC I decided it was a good idea since it did simplify the base class, and proceeded to fix up the child classes. Those changes provide a pointed demonstration of why inheriting from `AsyncSocket` is messy business. Nonetheless, most dependents also see some upside:

 - `TAsyncSocketIntercepted` gains `ReadVec` support.
 - `AsyncKTLSSocket` gains `ReadVec` support.
 - I identified an existing bug in `AsyncSSLSocket`'s `ReadVec` mode that causes loss of the correct error codes. Since it's not clear what to do (see inline comment), I kept its "current behavior" and will let the code owners decide what to do (if anything).

Reviewed By: simpkins, dmm-fb

Differential Revision: D43642341

fbshipit-source-id: e9614cdcf0f98caf29a688709e85bea2baf82728