Make keyset non-pessimised builtin
Summary:
Now that `keyset<T>` has only an array key bound on the generic parameter, we can make it a non-pessimised builtin. That is, `keyset<int>` is a keyset whose elements have type `int`, and `keyset<(~int & arraykey)>` is a keyset whose elements have type `(~int & arraykey)`. (Note the need for intersecting the dynamic bit of the type with `arraykey` to satisfy the bound.)

* Construction of keysets is changed: `keyset [ getLikeInt() ]` now produces `keyset<(~int & arraykey)>`, and `keyset [3]` produces `keyset<int>`.
* Indexing directly on keysets (not through an interface) gives back the actual element type e.g. if `$x` has type `keyset<int>` then `$x[0]` has type `int`.

Differential Revision: D44170654

fbshipit-source-id: 6d46ff85395b2a0e99db24087c68d75df1302ae6