Implement temporary flag project_metadata_with_flag_www_ready
Summary:
This flag is temporary to allow the safe rollout of the new project metadata format. It shall be removed in about 3 months, since 3-month old www revision are barely supported.

It's a flag in .hhconfig controlling whether the new project metadata format is supported for a specific www revision.

See new format rollout plan in attached task.

Reviewed By: nt591

Differential Revision: D44219008

fbshipit-source-id: 588c04d1978946e0157afd2f5ed7363c3ba7faee