remove configurability of default-max-bucket-size
Summary:
`--config max_bucket_size=500`.

Max bucket size is a setting in hh.conf. It's overridden by a JK. For the past year we have had the JK at a constant "500" for every machine in the company.

I think it is an ill-conceived setting. It is the "max bucket size" used for every place in the code that calls MultiWorker but fails to set its own bucket size. That's a weird idea. How can you usefully do tuning of this single parameter when there are so many different places that use MultiWorker, each with their own characteristics?

Also it's badly positioned in code, initialized in ServerEnvBuild.make_genv, which is a weird place to set up global state, and called by some tools but not by others depending on their needs.

In this diff I'm just deleting it. If we ever wish to use something similar in future, I think we can come up with a better control.

Reviewed By: nt591

Differential Revision: D44237592

fbshipit-source-id: 2bc9236d14999cde9f45e862a962b266adce0f34