ServerEnv tracks latest watchman clock
Summary:
ServerEnv.env is the structure which stores fields like "changed files which need to be re-indexed and then have their fanout typechecked", and "files which need to be re-typechecked".

This diff augments ServerEnv.env to also include the *watchclock*, the watchman clock at which we last got more changed files and stuck them into env.

I'm doing this now because of consistency-correctness with streaming errors. The desired end-user experience is that when a user types "hh" then they get a typecheck that reflects the state of disk no earlier than the moment at which they typed "hh". Therefore, hh_client consume a streaming errors.bin file, it needs to (1) know the watchclock corresponding to the state of the world at the time when the typecheck started, (2) judge whether any files have changed since then (which would render the errors.bin unusable).

This diff lays the foundation for part 1. Once watchclock is stored in env thanks to this diff, then it will be readily available to write at the start of errors.bin

Reviewed By: nt591

Differential Revision: D44243941

fbshipit-source-id: fb253f13c8f5cf68d15c70e5f254d4a5459cabd2