oops too many watchclock log messages
Summary:
I recently landed D44243941. However, it inadvertently spews messages to the server log once a second.

That's because recheck-loop runs once a second, and does `'Skip` (no-op) watchman checks on half of them, and the skip ones get back the watchclock `None` while the others get back a real `Some clock`. So, the code printed every time it changed between None and Some. This diff fixes it.

Reviewed By: nt591

Differential Revision: D44391788

fbshipit-source-id: 4f964fa151e85fbb3688fc15433cac7510c42b82