xrefs for modules
Summary:
A simple change overall but it is slightly annoying that it breaks the convention that decls were indexed in `Symbol_index_decls` and xrefs were indexed in `Symbol_index_xrefs`.  Reason is that module references aren't in the "occurrence lists" unlike the other references, but in the tast as a `set_module` node.

Maybe it could have been better to directly change the way we compute the list of occurrences.

Reviewed By: donsbot

Differential Revision: D44278909

fbshipit-source-id: 78ad249ee118749c60993822008facb59ec24095