status: show watchman crawl progress bar in rust
Summary:
While we are waiting for watchman to crawl the repo, call "debug-root-status" and pipe the file count into a progress bar.

The similar Python logic renders the "crawling" progress bar until the crawl is done and then proceeds with the query. In Rust I made it render the progress bar concurrently with the query. This should show the progress bar even if the repo has never been watched before (i.e. when there is no crawl until the query happens).

Reviewed By: quark-zju

Differential Revision: D44033615

fbshipit-source-id: 3c97fa7a63cd090b5ace82d7385f887e564eb9cf