Add HSL function completion on value types with -> syntax
Summary:
If a user writes `$item->` in the IDE and `$item` is a `vec`, `keyset`, `dict` or `string`, offer any HSL functions that are compatible.

This 'fake arrow completion' really helps with discovery of built-in functions on value types, which have historically been harder to find than methods on classes.

Reviewed By: hgoldstein

Differential Revision: D44044277

fbshipit-source-id: 5d2854b9df9f25331d662cffe3a189ad4ffde80a