splat support
Summary:
"Splat" is a weird operation:
```
$c = vec[1, 2, 3];
foo($a, $b, ...$c);
```
pretends that we called `foo` with the array values of `$c` unwrapped:
```
foo($a, $b, 1, 2, 3);
```
This has to be done at runtime because `$c` is a runtime value.  For now we'll just let Infer know about it by marking it as:
```
foo($a, $b, __sil_splat($c));
```

Also added support for the ResolveMethCaller instruction.

Reviewed By: artempyanykh

Differential Revision: D43666375

fbshipit-source-id: 4942eeb3e4a4cfce668be9189e6cb0353b16be31