atomically_create_and_init_file
Summary:
For the streaming errors file, I need a way to atomically create it with contents. (That's because server process creates it, and client process reads it, and client process should never see it with empty contents).

This diff implements atomic creation in the standard linux way, with O_TMPFILE.

Reviewed By: CatherineGasnier

Differential Revision: D44388834

fbshipit-source-id: b288466045498bea7515fe878032ccc20b133dc2