factor out post_watchman_filter
Summary:
This is a refactoring diff. It exposes `FindUtils.post_watchman_filter`.

(1) The deal is that we set up our watchman subscription with `FilesToIgnore.watchman_server_expression_terms`. This is a json expression, which we pass to watchman, which tells it to look for *.php and ignore .hg/* and so on.

(2) When watchman delivers us a list of changed files, we run it through `FindUtils.post_watchman_filter` which strips out files that match .hhconfig's `paths_to_ignore=` directive.

(3) Next, we check if two key files have changed (.hhconfig and PACKAGES.toml) and if so then hh_server must exit.

Previously, steps (2) and (3) were combined inside ServerMain.ml. This diff extracts step (2) out into a separate library so other parties can call it too (for instance, in a future diff, hh_client when it's doing streaming errors).

Why is there a split between the lower-level module `FilesToIgnore` and `FindUtils` which sits on top of it? I don't know, and I think they'd be cleaner if combined. But I won't do that right now; and `post_watchman_filter` uses FindUtils and hence has to go in the higher-level module.

Reviewed By: CatherineGasnier

Differential Revision: D44391908

fbshipit-source-id: df9a8de1ca8e4ac322c94cbf25aa5a4b8f9b70b4