Avoid walking stack frames in create_special_implicit_context
Summary:
Under certain circumstances, create_special_implicit_context needs to
walk up the stack to find the caller's func. This can be extremely
expensive.

Fix this by providing a variant of the builtin that explicitly takes
the Func*. Add a new bytecode, CreateSpecialImplicitContext, and use
that instead of calling the builtin directly. HHBBC will attempt to
convert calls to the builtin to the bytecode. Since the bytecode is
run directly in the caller, it can easily provide the Func* to the
runtime.

Having this bytecode also allows for future optimizations and
specializations.

Reviewed By: jano

Differential Revision: D44409567

fbshipit-source-id: c9859b402a00cb986daa0dac4b7961661fdef372