Allow `assemble_instr` to be public
Summary:
Allow assembly of a single bytecode from a string.

**Internal**:
The PIVV team is building a dynamic analysis tool for HHVM. As dynamic analysis follows the runtime execution of bytecodes, we need to way to extract instructions from a trace. Instead of re-inventing the wheel, we would like to leverage the existing assembler to extract single bytecodes from a program execution.

Reviewed By: aorenste

Differential Revision: D44437814

fbshipit-source-id: dbab8cab0fe07be867b946fb74c223ef8546699e