Fix entry state calculation for unreachable functions
Summary:
If a function is totally unreachable (for example, if it has a
parameter type constraint which is unsatisfiable), we were properly
marking the entry block as unreachable. However we were still pushing
the entry block onto the worklist. For the most part, this worked
because we'd process the first bytecode, realize we're in an
unreachable block, then bail. However, if that first bytecode can
possibly raise and push state to a catch block, we'll push that state
before bailing out. Since we pushed state, we'll mark the catch block
as reachable and then process that. However this can cause all sorts
of assertions because lots of things are going to be Bottom.

Don't push the entry blocks onto the worklist if its unreachable. This
requires distinguishing the main entry point and DV-inits (one may be
reachable and the other isn't).

Reviewed By: jano, hgoldstein

Differential Revision: D44491559

fbshipit-source-id: 76a7fd1867a6b2386c09703227c282fa62105053