monitor_update telemetry
Summary:
I want more telemetry about monitor_update. Which of the many cases it examines, actually occur in practice? (can we get rid of checks for the sql case?)

I'm doing this because my next project will be to do away with much of the informant and monitor, and switch to a simpler system where the sole basis for decision is when serverMain does a query on watchman. So I want to learn more.

There are two possible concerns with this diff:
1. Are we doing HackEventLogger on a critical path? Comments in the code warn against this. I don't think so: the critical path is only `Move_along` with no `reason`, and this diff doesn't add any logging in this case.
2. Will we end up with busy loops where there's a common event which is hit too many times? That's why I did throttling on the HackEventLogger side, to avoid this worry.

Reviewed By: CatherineGasnier

Differential Revision: D43996749

fbshipit-source-id: b7e52283f7af777aaa357fa78294fbd253a14c03