Keep track of et_hint inside Typing_env_types.in_expr_tree
Summary: When type checking an expression tree, it is useful to know the domain we are checking. This diff threads this information in the environment by passing in the et_hint portion of an ExpressionTree node. This will be used in future diffs to provide better error messages that can refer to the domain instead of expression trees in general.

Reviewed By: Wilfred

Differential Revision: D44384604

fbshipit-source-id: 2583fb718a3c501d86cd24ca30e245e6a22f0ad1