1/3 add ServerProgress.Errors library
Summary:
This diff adds an API, `ServerProgress.Errors`, for the streaming error file. This is its design:
```
OVERVIEW OF STREAMING ERRORS FILE

Errors is a file in /tmp/hh_server/<repo>/errors.bin which is written
by server during a typecheck. The file is first created when the server
starts its first typecheck, gets errors appended to it as the typecheck
discovers them, is unlinked and then another one with the same name
created each time the server starts another typecheck, and the file is finally
unlinked upon server exit. A "typecheck" can be thought of as a typecheck of the
entire program, and hence the errors file will contain every error reported in WWW.
But in reality the server does tricks to typecheck a smaller number of files just
so long as it still manages to report every error into the file.

Therefore: a client can at any time open an file descriptor to this file,
and "tail -f -n +1" (i.e. read from the start and then follow) to follow
the typecheck that is currently underway or most recently completed, from its
start to its end. It will still be able to read the file descriptor through
to its end even if, in the meantime, a new typecheck has started and the old
file has been unlinked. Note that a given file-descriptor will only ever point
to a file which monitonically grows.

The server can have its typecheck interrupted. Some interruptions like watchman
will cause the current typecheck to be cancelled, then a new typecheck started.
In this case the existing errors file will be unlinked, a new errors file created
for the new typecheck, and a sentinel will be written into the old file descriptor
to show that it ended before completing. Other interruptions like from
"hh --type-at-pos" will have no effect: the current typecheck file can continue
being read just fine. If a client reads a file descriptor on its sentinel, that is
a guarantee that a new errors file is already in place and can be opened (or,
if no errors file exists, that can only be because the server has terminated).

If file does not exist, the client knows that either there is no server, or there
is one but it has not yet started typechecking (it might still be loading saved state).
In either case it is appropriate for the client to clientConnect {autostart_server=true}
in the normal way, until it establishes a connection to the server, and then it can
wait for the errors file.

SEMANTICS OF ERRORS-FILE CONTENTS

* An errors-file is tied to a particular watchman clock value. It reflects
  all file-changes that happened prior to the clock. (As to whether it reflects
  any file-changes that happened after the clock, that is impossible to tell.)
* As mentioned above, when the errors-file is complete, it contains the full
  set of errors that hh_server knows about for the project.
* Each error report is an Errors.error list Relative_path.map, i.e. it encompasses
  several files, and for each file it has a list of errors. Currently we make
  one report for all "duplicate name" errors across the project if any, followed by
  one report per batch that had errors. This means that a file might be mentioned
  twice, once in a "duplicate name" report, once later. This will change in future
  so that each file is reported only once.
* It currently does not write empty reports, though that might change in future
  (e.g. we might decide to write "no errors in file a.php" for a file which
  previously did have errors).
* Within a single report, the error list has been sorted and de-duped.
```

There are a few design decisions not mentioned in the doc-comments which I'll note here...
1. I decided to stick this into the existing `ServerProgress.ml` rather than making a new file or new buck target. That's because it feels like progress.json and errors.bin are so very similar.
   1. However, ServerProgressLwt.ml has to be a separate module, because hh_server isn't allowed to link to Lwt
   2. We have an existing convention that "server/monitorConnection" is a low-level module that combines RPC types, writing and reading functions. Similarly, "server/serverProgress" is a module that combines RPC types, writing and reading functions.
2. I decided to make all the functions global. That reflects the fact that `root` is already global in both the server and client binaries (e.g. through Relative_path.set_prefix). It also reflects the fact that only one instance of hh_server can even exist for a given repository. That's as global as you can get!
3. We have to deal with what happens when a client connects to an errors.bin that was written by a later or earlier version of hh_server. I copied our existing pattern from monitorConnection, of having a loosely typed field `extra : Hh_json.json`, to allow for forward and backward compatibility without breaking the serialization format.
4. `ignore_hh_version` is currently controlled by the server, and I preserved that behavior. Thus for instance `hh start --ignore-hh-version` will allow all versions of the client to connect to it. (But using the flag in the client doesn't allow it to connect to arbitrary versions of the server).

## How to review this diff
1. Although it says "1300 lines changed in this diff", 650 of them are tests and 100 of them are docblocks, so it's really only 550 lines...
2. Read the big block of text in the summary
3. Read the `test_completed` and `test_async_read_completed ` unit tests. They show straightforward success cases that illustrate the API for consuming errors.
4. Focus on the producing side now. Read the ServerProgress.mli docblock on `write_state`, and docblocks for the other producing APIs in this module `new_empty_file`, `report`, `complete`, `unlink`. Now read the unit test case `test_produce_disordered`. This shows the "state-machine model" for producing errors.
5. Read ServerProgress.ml for the implementation. From a concurrency correctness perspective, focus on `with_lock`, `write_message`, `read_message`, `unlink_sentinel_close`. The rest of the methods are straightforward.
6. Read SeverProgressLwt.{mli,ml}

Reviewed By: CatherineGasnier

Differential Revision: D44159582

fbshipit-source-id: d469c91de267947fa490d3d47e152843439c97c3