Set pessimise_bulitins=true by default
Summary:
Now that we don't use `pessimise_builtins` for `vec`, `dict`, `keyset`, or `Pair`, we only care about the flag for mutable collection types. To simplify the matrix of options, let's ignore the flag and just assume it is true.

A subsequent diff will remove the flag.

Some of the tests that specifically tested against the flag being false are now changed and look a little odd.

Reviewed By: vassilmladenov

Differential Revision: D44511269

fbshipit-source-id: 95ad72005eedbfa5c8481853f32696156a3015f4