Infer more precise types for collection constructors
Summary:
Previous diff D44091378 had a bug that affected nested constructors with explicit types, for example
```
vec<vec<int>>[vec[2,3]]
```
would be assigned a type `vec<vec<~int>>`.

This was because we pass an expected type of `~<vec<int>>` down when inferring a type for the element of the outer vector, and then simply check against this type and use it as the inferred type.

The fix is to use the pessimised type as the expected type, but then check the result first against the non-pessimised type, and if this succeeds, to use this as the inferred type.

I've also taken the opportunity to clean up the code somewhat, though the logic is still very complex!

Note: some of our errors change a little (in fact I think they revert to what they were before the previous diff).

Differential Revision: D44460678

fbshipit-source-id: bc3196ec6dc7a02eb4e736c97a7de1cbba01d90e