Add heuristic for like-subtyping
Summary: If we have a `newtype T<T1> as f(T1)` then subtyping `T<t> <: ~T<u>` should just consider `T<t> <: T<u>` and not try to expand the bound on `T<t>` to `f(t) <: T<u>` or `f(t) <: dynamic`. The first isn't helpful unless `T<u>` has a lower bound. While it is possible for the second to be necessary, in practice the extra disjunction seems to get in the way far more often then it helps.

Reviewed By: andrewjkennedy

Differential Revision: D44543725

fbshipit-source-id: 6e0cf1eff7c913b8fcb4a2c7c493bbdfe28f969c