Fine-grained implementation of function call on supportdyn functions, avoiding exponential
Summary:
The cost of inferring method invocation on `supportdyn` functions is potentially exponential in the size of the expression. Consider
```
<<__SupportDynamicType>>
class C {
  public function foo(C $c):C {
    return $c;
  }
}
function getVec():vec<~C> {
  return vec[];
}
function testit():void {
  $lc = getVec()[0];
  $c = new C();
  $c->foo($c->foo($c->foo($c->foo($c->foo($lc)))));
}
```
To check the outer call to `foo`, we use our SDT function call logic, which first attempts to check the call "statically" i.e. with type `C->C`, and then checks the call "dynamically", which it does by checking that all parameters both support dynamic and are subtypes of `~C`. But rather than inferring the type of the argument before doing this work, it passes the argument expression down to the call logic so that bidirectional type checking can be used i.e. an expected type passed in to inference. This means that the argument expression potentially gets checked twice. In the worst case, when all static checks fail, we end up with O(2^n) cost for an expression (such as in the example above) nested n deep.

Instead of this repeated check on *expressions*, this diff checks the argument expressions only once, with like-types as expected types. It then goes through each parameter one-by-one, checking to see if the inferred type is a subtype of the non-like-type i.e. normal static calling. If the check fails, it checks against the like-type. It accumulates a boolean (`used_dynamic`) that is set to true for any argument that fails the static check. If `used_dynamic` turns out to be true, it checks that all the argument types (even those that succeeded without a like type) support dynamic, and finally adds a like-type to the return type of the function - i.e. implementing `supportdyn` function calling.

The new approach is not complete, and it fails to infer a good type where the old technique succeeded. Here is a simple example:
```
<<__SupportDynamicType>>
function iter<Tv1 as supportdyn<mixed>>(vec<Tv1> $traversable, supportdyn<(function (Tv1): void)> $value_func): void {
}

<<__SupportDynamicType>>
function print_line(string $_):void { }

function testit(vec<~string> $vs1, ~vec<string> $vs2):void {
  $f = print_line<>;
  // Works, because we infer Tv=string
  iter($vs2, $f);
  // Works, because we're explicit
  iter<string>($vs1, $f);
  // Does not work, because we commit to Tv=~string
  iter($vs1, $f);
}
```
Consider the final call to `iter`. We first attempt to check `vec<~string>` against `vec<#1>` where `#1` is the type variable generated for `Tv1`. This succeeds, with constraint `~string <: #1`. We then proceed to check `(function(string):void)` against `(function(#1):void)`, and deduce that `#1 <: string`. Clearly this is inconsistent with the previous constraint, so the subtype check fails. We will then try the same argument type against a pessimised parameter type, which also fails. If instead, on the first argument we had tried the "dynamic" check of `vec<~string>` against `~vec<#1>`, we would have succeeded (with like-pushing), generating constraint `string <: #1`, which would then have allowed the second parameter to type-check. Clearly if we tried 2^n combinations of argument against non-like/like parameter type, we would have accepted this program.

For now, we will accept this incompleteness, but consider further improvements in future.

Differential Revision: D44294183

fbshipit-source-id: 4882c339432d7bf6feed1a8444976883aa89aa60