Eval Virtualized Expressions in Expression Trees with Default Capabilities
Summary:
# Context
When type checking expression trees we relied on the fact that we virtualized expressions to lambda expression to avoid managing the local environment directly. This work for most cases, except for the locals used to track the capabilities of a function. Lambdas automatically take on the capabilities.

This leads to reporting unnecessary errors when type checking expression trees in function that do not use the default set of capabilities. An example of this would be

```

async function bar(
  ExampleContext $_,
): Awaitable<ExampleDslExpression<ExampleFunction<(function(): ExampleInt)>>> {
  throw new Exception();
}

function pure()[]: void {
  // Reported an error that we couldn't call `bar()` due to incompatible capabilities
  ExampleDsl`bar()`;
}
```
Here we reported an error that invoking `bar()` within the pure function was not allowed because `ExampleFunction<(function(): ExampleInt)>` is a function with default capabilities. However the method `bar` isn't ever invoked, it is only virtualized as a call for type checking purposes. This has real impact on the usability of expression trees because helper functions written for the DSL will have to deal with effect polymorphism.

# Mitigation
To fix this I skipped coeffect checks if `Typing_env.is_in_expr_tree` is true. This needs to be done in two locations.

First in `Typing_local_ops.check_local_capability` and second inside `Typing.call` inside `check_implicit_args`.

Reviewed By: losvald

Differential Revision: D44563181

fbshipit-source-id: d5f1c6bb92c00c4b630f93b72737c64a588891c2