refactor: simpler traversal logic
Summary:
The intent is to do the same thing in a way that is easier to read and avoids repeat traversals

I updated one of the snapshot tests to use a method to make sure the logic for traversing methods is exercised. Existing tests with `_var` and `_prop` in their names exercise the logic for what we don't want to traverse.

Reviewed By: Wilfred

Differential Revision: D44497208

fbshipit-source-id: 124c351a087df0c0f9dc554678980df46147e5a4