server failed to update clock
Summary:
History:
1. In D44243941 I made ServerEnv.env keep track of watchclock. But this was purely informational: nothing yet consumed the watchclock, so I had no good way to check that it was correct.
2. I tweaked it a little further in D44391788.
3. The first thing to consume the watchclock was D44159580, "cli streaming errors". But that had a bug in it which meant it didn't read the watchclock correctly.
4. I fixed that bug in D44607507 and now for the first time am exposed to situations where the watchclock isn't updated correctly...

Very quickly I hit a bug. Repro: I modified a file on disk, and then a subsequent `hh --config ide_standalone=true` just waited indefinitely for "hh_server sync". Here are the server logs:
```
[2023-04-02 07:32:02.048] [master] SERVER_IS_READY. Took 2.411028 seconds to init.
[2023-04-02 07:32:12.322] [master] Watchman response: {"subscription":"hh_type_check_watcher.2493203","root":"/data/users/ljw/www","clock":"c:1680373270:49066:1:9740","files":["flib/intern/entity/employee/EntEmployee.php"],"is_fresh_instance":false,"unilateral":true,"since":"c:1680373270:49066:1:9738","version":"2023-03-13T02:08:08Z"}
[2023-04-02 07:32:12.322] [master] Errors-file: starting new fd31:215265803
[2023-04-02 07:32:12.322] [master] ******************************************
[2023-04-02 07:32:12.322] [master] Check kind: will bring hh_server to consistency with code changes, by checking whatever fanout is needed ('Full_check')
[2023-04-02 07:32:12.322] [master] Processing changes to 1 file: /data/users/ljw/www/flib/intern/entity/employee/EntEmployee.php
```

Conspicuously the logs do NOT contain the line
```
Recheck at watchclock c:1680373270:49066:1:9740
```

The deal is that `query_kind = 'Skip` doesn't skip watchman updates as you might think. Instead, as far as I can tell, it's identical to `query_kind = 'Async`, i.e. picking up any pending watchman updates that have been send over the subscription. (I think this behavior was an unintended consequence of D8341904). Anyway, my code which decided to not update `env.clock` in case of `'Skip` was just wrong.

This code is rough and in need of a rewrite, which I intend to do in the future T146551552. But for now, this diff merely picks up the clock correctly, and adds copious comments.

Reviewed By: CatherineGasnier

Differential Revision: D44614181

fbshipit-source-id: ec68004ad5f1fc04b796cb3d07c8146457fe1a2f