Improve string generation in Python
Summary:
Most of the logic in `value:py3_string_value` is dead because the quotes are removed during parsing. It also doesn't escape special characters. Instead of trying to fix it, replace with generic `value:string_value` which works fine with Python. This simplifies the codegen and ensures that string literals are consistent with C++ and other languages which use `value:string_value`.

Given the following Thrift string literals:

```
  "\x01",
  "\x1f",
  "\x20", // space
```

They are emitted as follows before and after the change.

Before:

```
" ", " ", " " // non-printable characters replaced with spaces to avoid breaking version control but they are rendered the same
```

After:

```
"\001", "\037", " "
```

Notice that control characters where not escaped previously leading to confusion.

Similarly, update the legacy Python generator to escape control characters.

Reviewed By: iahs

Differential Revision: D44547137

fbshipit-source-id: fb9add317f3baa55fa20b348e0e27888e8f676c7