@cpp.Type
Summary:
Because of the proliferation of plugins and other bespoke usages of the AST library in C++, we have to double-write annotations as both structured and unstructured during the migration. Instead of doing this in the IDL which is ugly and prone to desyncing, implement a lowering mechanism and use it to support the structured annotation in the compiler. Once the migration is complete we can change our callsites to just look for the structured annotation; until then looking for both doesn't add value.

Since cpp.type and cpp.template are closely related they are combined into a single structured annotation with a validator to enforce correct usage.

Reviewed By: vitaut

Differential Revision: D44561471

fbshipit-source-id: 6f1adbe5b2d94c69cd15d654c651b100ec2a03c2