Store bytecode separately while using extern-worker in HHBBC
Summary:
Right now the bytecode for a given php::Func is stored with it, and
serialized along with it. However, in some situations we may want to
have a php::Func available, but not need the bytecode. It's more
efficient in this case not to provide the extern-worker job with the
bytecode. We can accomplish this by storing it separately and only
providing it when explicitly needed.

php::Func will continue to have a pointer to it's bytecode. However
the serialization will not read it (will assert if non-null). Instead
the bytecode is serialized in a separate FuncBytecode struct. Inside
an extern-worker job, the bytecode (if provided) can be manually
stored in the php::Func.

Reviewed By: kavoor

Differential Revision: D44618744

fbshipit-source-id: 880002ec89a83281e4d72b5767eb094d5cb38d3a