Rename extractMaskFromPatch to extractMaskViewFromPatch
Summary:
Rename to `extractMaskViewFromPatch` which currently stores the address of Value. This is required for us to operate map key Value which is not integer or string. Therefore, if the extracts mask view is never crossed over the wire, users can use the `extractMaskViewFromPatch` to extract read and write mask. This is also effectively more efficient that storing string as it requires to copy the string key. Therefore, change the name to `MaskView`.

We should consider removing `extractMaskVIewFromPatch` since all current use cases is covered with string and integer and if there is no demand in the future. This MaskView cannot be send over the wire, and the user has to be extract cautious about the lifetime.

Reviewed By: rahku

Differential Revision: D44481286

fbshipit-source-id: 1f658eec1b7d1d9b4acd71101ad2c619b7270a3c