Use Contextual Information When Reporting ExprTree Splicing Errors
Summary:
## Context

We received reports from users that the errors we report when an incorrect value is spliced into an expression tree is confusing. I believe this because our errors refer to generic concepts like "Spliceable" and "Expression Tree" when most users interact with the DSL that uses expression trees as an implementation detail.

To make these errors more comprehensible we can include contextual information in our error reports that references positions and types within the DSL instead of the more generic expression tree terminology. We have overridden errors this way already in other places such as when we replace missing method errors with more specific errors when a DSL doesn't support certain virtualized features.

## Solution

I utilize the additional context of the `dsl` that was added in {D44384604}. I use this to change the error reported when a non-Spliceable value is attempted to be spliced into an expression tree. The general structure is:

```
This value cannot be spliced into a `{DSL}` expression tree (Typing[4110])
  Expected `{DSL Spliceable}` because you are splicing into a `{DSL}` block
  But got `{Expr Ty}`
  Hack values need to be lifted to compatible types before splicing.

See {Docs URL} for more information.
```
Here `{DSL}` is the name of the DSL implemented using expression trees, namely for an expression tree:
```
ExampleDSL`<expr>`
```
it will be the string `ExampleDSL`

For `{DSL Spliceable}` we pull the type of the third argument to the `splice` method on the DSL class. This can sometimes be a more specific type than `Spliceable`. At the very least the position reported will be to the `splice` method where additional doc comments can be supplied to detail how values should be lifted into the DSL

`{Expr Ty}` is the type of the expression that is being spliced

Finally `{Docs URL}` will refer to the string contained in the `<<__Docs(<string>)>>` associated with the DSL class or to the docs on splicing for the Hack manual.

Reviewed By: Wilfred

Differential Revision: D44517900

fbshipit-source-id: 06246363cd80e04a22a2b0a67406a716095c7cee