Simplify dependency graph code
Summary:
Separating adapter and struct dependencies makes the code much more complicated, and misses cases including typedef-of-container-of-typedef. Combine them, add a test for the missing case, and fix revealed issues.
Sadly one of the revealed issues is that frozen always requires complete types for containers, but frozen is enabled for so many structs that we run into false circular dependencies if we start tracking these dependencies whenever frozen is enabled. Instead we add `cpp.container_require_complete_params` to manually opt in the handful of containers that are actually running into this problem.

Reviewed By: vitaut

Differential Revision: D44521696

fbshipit-source-id: 3c13e17bccc62ea9506da21ee9a6965264a5cb84