plumbing before adding actual naming table full index build
Summary:
This diff introduces the new flag, state, and notifications to the daemon for the full index naming table build. The feature is gated behind the `.hhconfig` flag `ide_fall_back_to_full_index`

The daemon enters the `full_index_fallback` state when it fails to load a naming table (failing to download it or loading one that's provided). In this diff, all this does is wait 60 seconds and then return the error from failing to load the naming table. The next diff plugs in the FFI.

Reviewed By: ljw1004

Differential Revision: D43677974

fbshipit-source-id: 6ad2a8765e08ec9b1ecde0e5ac7ec324687e24e7