Prefer errors from constraint types
Summary:
When we report errors on disjunctive subtyping queries, we pick the first error to report to the user. This causes unintuitive errors, but in the case of nullsafe property access when the property is not available/or is hidden due to private/protected visibility, the error is complete illegible, e.g.,

```
Typing error (Typing[4110])
  File "A.php", line 13, characters 3-9:
  Expected `null` (use of `?->` operator)
  File "A.php", line 6, characters 26-32:
  But got `A`
```

This happens because nullsafe operator is typechecked by weakening the `Thas_member`  constraint by unioning it with a `null` before doing the subtype query.

This diff reorders the disjunctive query so that we check against `Thas_member` first and consequently report on its error if the whole query fails. This is a heuristic but it works well for this case and doesn't regress any other case

Reviewed By: andrewjkennedy

Differential Revision: D44629889

fbshipit-source-id: 1307e4f98ea4b3c2759ef320105cfba1354f229f