Assert type of the array emitted in NewVec
Summary:
Refine the type of a newly emitted vec based on the number of elements that are initialized.

Note that hhbbc already emits AssertRATs for arrays that are created in Hack. However, there are many places in the jit where we emit new vecs, for example when passing variadic args to a function (https://fburl.com/code/icnwtcyk). The newly added Assert in this diff will provide a more specific type for such arrays.

Also, it isn't great that we're creating RepoAuthTypes/RepoAuthType::Arrays in the jit. But unfortunately, this is what the jit type uses to store array specialization. In the long term, we may want to look into introducing a newer type that can represent array specialization information in the jit which can help us preserve the types after array operations like Append, which currently pessimizes the type to a TVec with no specialization. The newly introduced type can be short-lived for the duration of the translation, as opposed to a RepoAuthType::Array that needs to be alive for the entire duration in case new code needs to be jit-ed in the future.

The element types are fairly pessimistic as well, for similar reasons as stated above. The cost associated with creating RepoAuthType::Arrays for all possible combinations of elements can be really high as they are stored in memory. Moreover, there isn't a great way to convert the jit/type available for each element to RepoAuthType, especially for unions. Since the most accurate type isn't a strict requirement for the purposes of the current change, the element types have been pessimized to InitCell. Refining the type can provide other wins too by exposing optimization opportunities to other passes in the jit.

Reviewed By: ricklavoie, jano

Differential Revision: D44473614

fbshipit-source-id: 3cab6c4964331a30546d061918950c660df50c85