Optimize Str\format to ConcatStr if format string is trivial
Summary:
A vast majority of calls to Str\format in www contain either 1 or 2 string tokens. But we still call the built-in function vsprintf_l to get the final output string. There are several overheads in doing so like calling a c++ function, unnecessary creation/handling of vecs to pass in the arguments as well as handling for locales and other format specifiers.

We could potentially expand this to handle %d as well which is the next most common specifier.

Reviewed By: jano

Differential Revision: D44049156

fbshipit-source-id: a0364806709b444826edf2deae3ef0bc97e28834