Suggest splicing in variables for Expression Trees
Summary:
## Context

Users of expression trees may forget to splice in a variable that was defined outside of the expression tree. Currently the error reporting isn't very helpful and is the generic "Variable %s is undefined, or not always defined" error. We can provide a more contextual error stating that the variable is undefined in the expression tree, but present in the outer scope. We can then provide a quickfix to splice in the variable.

## Solution

I update `Typing_env_types.expr_tree_env` to include the locals from the outer environment when we enter it. I then updated the error we report when attempting to access an undefined variable. If we are in an expression tree, we will include the set of locals in the outer environment when deciding what variable is the most similar.

Ordering is relevant here, we want to have the suggestions that are already present in the expression tree to be preferred. I also added a specific `Naming_error` entry for reporting this new error when the most similar suggestions is in the outer scope. The quickfix will then simply splice in the variable that is suggested.

Reviewed By: Wilfred

Differential Revision: D44531628

fbshipit-source-id: f56ee048f733e7897039e2806444e0da8705b747