full index fall back
Summary:
Add the ability to fall back to performing a full index naming table build of the repository if loading a saved state fails. This functionality is gated behind the flag `ide_fall_back_to_full_index`. Add the line
```
ide_fall_back_to_full_index = true
```
in the `.hhconfig` file in the root of the repository to enable this feature.

Reviewed By: ljw1004

Differential Revision: D43927284

fbshipit-source-id: 1f5d905768e432eca2bdba4b72b0af5fc2fe8d71