Move property initial value rewriting in HHBBC into extern-worker
Summary:
HHBBC will rewrite the initial values of properties in order to infer
better types. Move this logic over to use extern-worker. It can be
done as part of FlattenJob and BuildSubclassListJob. This means that
BuildSubclassListJob now needs php::Class and not just ClassInfo. It
does not, however, require the bytecode, which lessens the overhead.

Since BuildSubclassListJob now have the php::Class, we no longer need
to mirror the class Attr bits.

Reviewed By: kavoor

Differential Revision: D44649907

fbshipit-source-id: 06605b24ca2eae2575d8e1a8e2f06809a6fc2863